requirements.txt
torch==2.3.1
torchvision==0.18.1

transformers==4.41.2
huggingface-hub==0.23.0

tqdm==4.66.4
pandas==2.2.2
numpy==1.26.4
Pillow==10.3.0
requests==2.32.3

faiss-cpu==1.8.0

Flask==3.0.3
gunicorn==22.0.0
sentencepiece
protobuf

python-dotenv
pinecone
